Struct sp_keyring::sr25519::sr25519::Public

pub struct Public(pub [u8; 32]);

An Schnorrkel/Ristretto x25519 (“sr25519”) public key.

Implementations

impl Public

pub fn from_raw(data: [u8; 32]) -> Public

A new instance from the given 32-byte data.

NOTE: No checking goes on to ensure this is a real public key. Only use it if you are certain that the array actually is a pubkey. GIGO!

pub fn from_h256(x: H256) -> Public

A new instance from an H256.

NOTE: No checking goes on to ensure this is a real public key. Only use it if you are certain that the array actually is a pubkey. GIGO!

pub fn as_array_ref(&self) -> &[u8; 32]

Return a slice filled with raw data.